Add remove book button
